# Project/ai-learning-hub/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Learning Hub API
  version: 0.1.0
  description: |
    当前包含健康检查和第一版认证契约。第一版仅支持邮箱/密码注册登录；不提供邮箱验证和密码找回。
servers:
  - url: /api/v1
paths:
  /auth/register:
    post:
      summary: 注册学习者账号
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: 邮箱已被注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: 使用邮箱和密码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: 邮箱或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: 刷新访问令牌
      operationId: refreshAccessToken
      description: 从 HttpOnly refresh cookie 读取令牌；该令牌不会出现在 JSON 响应中。
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 刷新令牌无效或已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: 登出并撤销当前刷新令牌
      operationId: logout
      description: 撤销当前 HttpOnly refresh cookie 对应的服务器端令牌记录。
      responses:
        '204':
          description: 已登出
  /me:
    get:
      summary: 获取当前登录用户
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      summary: 获取服务健康状态
      operationId: getHealth
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
          example: learner@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: LearningHub2026!
        displayName:
          type: string
          minLength: 1
          maxLength: 50
          example: 小明
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: 短期 JWT；前端仅保存在内存中。
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          description: access token 的有效秒数。
          example: 900
        user:
          $ref: '#/components/schemas/CurrentUser'
    CurrentUser:
      type: object
      required: [id, email, displayName, roles]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [LEARNER, CONTENT_ADMIN, SYSTEM_ADMIN]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
          example: UP
        service:
          type: string
          example: ai-learning-hub-api
  responses:
    ValidationError:
      description: 请求字段无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT